+++ kitty/options/definition.py
@@ -1349,10 +1349,17 @@ opt('hide_window_decorations', 'no',
     option_type='hide_window_decorations', ctype='uint',
     long_text='''
 Hide the window decorations (title-bar and window borders) with :code:`yes`. On
-macOS, :code:`titlebar-only` and :code:`titlebar-and-corners` can be used to only hide the titlebar and the rounded corners.
+macOS, :code:`titlebar-only` and :code:`titlebar-and-corners` can be used to
+only hide the titlebar and the rounded corners.
+
+On Wayland, :code:`titlebar-only` can be used to hide the titlebar while keeping
+the window shadow for resizing. On compositors that have server-side
+decorations (such as anything but GNOME), both :code:`yes` and :code:`titlebar-only` force
+client-side decoration mode.
+
 Whether this works and exactly what effect it has depends on the window manager/operating
 system. Note that the effects of changing this option when reloading config
-are undefined. When using :code:`titlebar-only`, it is useful to also set
+are undefined. When using :code:`titlebar-only` on macOS, it is useful to also set
 :opt:`window_margin_width` and :opt:`placement_strategy` to prevent the rounded
 corners from clipping text. Or use :code:`titlebar-and-corners`.
 '''